X86 64

+++ src/kernel/include/kernel/drivers/framebuffer/framebuffer.hpp
@@ -11,10 +11,10 @@ namespace kernel::drivers::framebuffer {
     inline constexpr std::uint32_t RGB_BLUE  = 0x000000FF;
     
     struct FrameBufferInfo {
-        std::uint32_t width;
-        std::uint32_t height;
-        std::uint32_t pitch;
-        std::uint8_t bpp;
+        std::uint64_t width;
+        std::uint64_t height;
+        std::uint64_t pitch;
+        std::uint16_t bpp;
         std::uint8_t* vram;
     };
